Refactor BattleField and BattleSnake; improve memory management and update method names for clarity

+++ Thanos/Enums/Constants.cs
@@ -5,17 +5,17 @@ public static class Constants
     public const int MaxSnakes = 8; 
     public const int CacheLineSize = 64;
     
-    public const int Empty = 0; 
-    public const int Me = 1;    
-    public const int Enemy1 = 2;
-    public const int Enemy2 = 3;
-    public const int Enemy3 = 4;
-    public const int Enemy4 = 5;
-    public const int Enemy5 = 6;
-    public const int Enemy6 = 7;
-    public const int Enemy7 = 8;
-    public const int Food = 127;
-    public const int Hazard = 255;
+    public const byte Empty = 0; 
+    public const byte Me = 1;    
+    public const byte Enemy1 = 2;
+    public const byte Enemy2 = 3;
+    public const byte Enemy3 = 4;
+    public const byte Enemy4 = 5;
+    public const byte Enemy5 = 6;
+    public const byte Enemy6 = 7;
+    public const byte Enemy7 = 8;
+    public const byte Food = 127;
+    public const byte Hazard = 255;
     
     public const int MaxBodyLength = 256;
 }
